+++ src/public/vault/borrow/execute/Deposit.sol
@@ -71,8 +71,8 @@ abstract contract Deposit is
                 deltaUserFutureRewardCollateral: deltaFuture.deltaUserFutureRewardCollateral,
                 deltaProtocolFutureRewardBorrow: deltaFuture.deltaProtocolFutureRewardBorrow,
                 deltaProtocolFutureRewardCollateral: deltaFuture.deltaProtocolFutureRewardCollateral,
-                blockNumber: block.number,
-                auctionStep: CommonMath.calculateAuctionStep(startAuction, block.number)
+                blockNumber: uint64(block.number),
+                auctionStep: CommonMath.calculateAuctionStep(startAuction, uint64(block.number))
             })
         );
