+++ src/public/vault/borrow/execute/Withdraw.sol
@@ -80,8 +80,8 @@ abstract contract Withdraw is
                 deltaUserFutureRewardCollateral: deltaFuture.deltaUserFutureRewardCollateral,
                 deltaProtocolFutureRewardBorrow: deltaFuture.deltaProtocolFutureRewardBorrow,
                 deltaProtocolFutureRewardCollateral: deltaFuture.deltaProtocolFutureRewardCollateral,
-                blockNumber: block.number,
-                auctionStep: CommonMath.calculateAuctionStep(startAuction, block.number)
+                blockNumber: uint64(block.number),
+                auctionStep: CommonMath.calculateAuctionStep(startAuction, uint64(block.number))
             })
         );
