+++ src/structs/state/AuctionState.sol
@@ -6,5 +6,5 @@ struct AuctionState {
     int256 futureCollateralAssets;
     int256 futureRewardBorrowAssets;
     int256 futureRewardCollateralAssets;
-    uint256 startAuction;
+    uint64 startAuction;
 }
